# Optimize GreenNode.CreateList