--- a/src/Compilers/Core/Portable/Syntax/GreenNode.cs
+++ b/src/Compilers/Core/Portable/Syntax/GreenNode.cs
@@ -937,7 +937,7 @@ namespace Microsoft.CodeAnalysis
                 }
             }
         }
-        
+
         public SyntaxNode CreateRed()
         {
             return CreateRed(null, 0);