--- a/src/Compilers/Core/Portable/Syntax/SyntaxNodeOrTokenList.cs
+++ b/src/Compilers/Core/Portable/Syntax/SyntaxNodeOrTokenList.cs
@@ -316,7 +316,7 @@ namespace Microsoft.CodeAnalysis
                 return default(SyntaxNodeOrTokenList);
             }
 
-            var newGreen = GreenNode.CreateList(items, n => n.UnderlyingNode!)!;
+            var newGreen = GreenNode.CreateList(items, static n => n.UnderlyingNode!)!;
             if (newGreen.IsToken)
             {
                 newGreen = Syntax.InternalSyntax.SyntaxList.List(new[] { newGreen });