--- a/src/Compilers/Core/Portable/Syntax/SyntaxToken.cs
+++ b/src/Compilers/Core/Portable/Syntax/SyntaxToken.cs
@@ -472,19 +472,26 @@ namespace Microsoft.CodeAnalysis
             return this.WithLeadingTrivia((IEnumerable<SyntaxTrivia>?)trivia);
         }
 
-        /// <summary>
-        /// Creates a new token from this token with the leading trivia specified.
-        /// </summary>
-        public SyntaxToken WithLeadingTrivia(IEnumerable<SyntaxTrivia>? trivia)
+        private static IReadOnlyList<SyntaxTrivia>? GetList(IEnumerable<SyntaxTrivia>? trivia)
         {
-            IReadOnlyList<SyntaxTrivia>? list = trivia switch
+            if (trivia is null)
+                return null;
+
+            return trivia switch
             {
                 null => null,
                 IReadOnlyList<SyntaxTrivia> l => l,
                 _ => trivia.ToArray()
             };
+        }
+        
+        /// <summary>
+        /// Creates a new token from this token with the leading trivia specified.
+        /// </summary>
+        public SyntaxToken WithLeadingTrivia(IEnumerable<SyntaxTrivia>? trivia)
+        {
             return Node != null
-                ? new SyntaxToken(null, Node.WithLeadingTrivia(GreenNode.CreateList(list, t => t.RequiredUnderlyingNode)), position: 0, index: 0)
+                ? new SyntaxToken(null, Node.WithLeadingTrivia(GreenNode.CreateList(GetList(trivia), t => t.RequiredUnderlyingNode)), position: 0, index: 0)
                 : default(SyntaxToken);
         }
 
@@ -509,14 +516,8 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public SyntaxToken WithTrailingTrivia(IEnumerable<SyntaxTrivia>? trivia)
         {
-            IReadOnlyList<SyntaxTrivia>? list = trivia switch
-            {
-                null => null,
-                IReadOnlyList<SyntaxTrivia> l => l,
-                _ => trivia.ToArray()
-            };
             return Node != null
-                ? new SyntaxToken(null, Node.WithTrailingTrivia(GreenNode.CreateList(list, t => t.RequiredUnderlyingNode)), position: 0, index: 0)
+                ? new SyntaxToken(null, Node.WithTrailingTrivia(GreenNode.CreateList(GetList(trivia), t => t.RequiredUnderlyingNode)), position: 0, index: 0)
                 : default(SyntaxToken);
         }