--- a/src/Compilers/Core/Portable/Syntax/SyntaxTokenList.cs
+++ b/src/Compilers/Core/Portable/Syntax/SyntaxTokenList.cs
@@ -361,7 +361,7 @@ namespace Microsoft.CodeAnalysis
                 return this;
             }
 
-            return new SyntaxTokenList(null, GreenNode.CreateList(list, n => n.RequiredNode), 0, 0);
+            return new SyntaxTokenList(null, GreenNode.CreateList(list, static n => n.RequiredNode), 0, 0);
         }
 
         /// <summary>
@@ -377,7 +377,7 @@ namespace Microsoft.CodeAnalysis
 
             var list = this.ToList();
             list.RemoveAt(index);
-            return new SyntaxTokenList(null, GreenNode.CreateList(list, n => n.Node!), 0, 0);
+            return new SyntaxTokenList(null, GreenNode.CreateList(list, static n => n.Node!), 0, 0);
         }
 
         /// <summary>
@@ -423,7 +423,7 @@ namespace Microsoft.CodeAnalysis
                 var list = this.ToList();
                 list.RemoveAt(index);
                 list.InsertRange(index, newTokens);
-                return new SyntaxTokenList(null, GreenNode.CreateList(list, n => n.Node!), 0, 0);
+                return new SyntaxTokenList(null, GreenNode.CreateList(list, static n => n.Node!), 0, 0);
             }
 
             throw new ArgumentOutOfRangeException(nameof(tokenInList));