This PR improves the performance of GreenNode.CreateList.
This is part of a larger goal to improve NormalizeWhitespace performance, the benchmark below tests NormalizeWhitespace performance of real-world generated C# source (~116k loc when normalized).
Benchmark:
|   Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 | Gen 2 | Allocated |
|--------- |--------:|---------:|---------:|-----------:|-----------:|------:|----------:|
| Baseline | 5.806 s | 0.0556 s | 0.0493 s | 74000.0000 | 18000.0000 |     - | 591.82 MB |
| New      | 5.595 s | 0.1072 s | 0.1731 s | 49000.0000 | 13000.0000 |     - | 393.68 MB |

Tested on Net5.0 RC2 with the Benchmark runtime being .NET Core 3.1.8

I also investigated a couple of other approaches (ImmutableArray + SelectAsArray, GetEnumerator directly) but this one came out on top.
Thanks for the help @333fred and @CyrusNajmabadi 😄 

Benchmark Files (drop into IdeCoreBenchmarks for testing):
[normalization-sample.txt](https://github.com/dotnet/roslyn/files/5373857/normalization-sample.txt)
[NormalizeWhitespaceBenchmark.cs](https://github.com/dotnet/roslyn/files/5373859/NormalizeWhitespaceBenchmark.txt)