@HurricanKai 2020-10-14
Here are some performance figures. The currently committed version is `List` previously it was `IROList` and the current master is `Baseline` the others are for other alternatives I've considered.
The numbers aren't 100% accurate but should give a good idea what's fastest. Sorted by speed.
|     Method |    Mean |    Error |   StdDev |      Gen 0 |      Gen 1 | Gen 2 | Allocated |
|----------- |--------:|---------:|---------:|-----------:|-----------:|------:|----------:|
| List       | 5.683 s | 0.0338 s | 0.0316 s | 72000.0000 | 18000.0000 |     - | 579.89 MB |
| List/Array | 5.861 s | 0.0505 s | 0.0472 s | 67000.0000 | 17000.0000 |     - | 536.25 MB |
| IROList    | 5.889 s | 0.0371 s | 0.0310 s | 49000.0000 | 13000.0000 |     - | 393.68 MB |
| Baseline   | 5.912 s | 0.0356 s | 0.0316 s | 74000.0000 | 18000.0000 |     - | 591.82 MB |