@CyrusNajmabadi 2020-10-14
Based on those numbers, looks like we should stick with IROL and not use List.  Or, we have an overload for both or something (but htat seems like a lot of complexity).