@333fred 2020-10-15
Is the `List<T>` version generating more garbage from the non-List paths? Looking at those numbers, while using IROL list a clear win in terms of memory, it doesn't actually make a different CPU-time-wise, as the IROL and Baseline are within the Error margin. I think I'd prefer to have both a List and IROL overload. The method complexity is not very high, and 200 MB of memory savings on a large `NormalizeWhitespace`, while nice, is basically peanuts for our scenarios.