@HurricanKai 2020-10-15
The only types every passed in are `List<T>` and `T[]` I've tested two overloads with those two, see `List/Array`, I'll test List/IROList too, but I'd be surprised if it yielded better/much different results. Maybe because most type passed to WithTrivia are IROList 🤔