@HurricanKai 2020-10-20
I've just tested this but the results aren't beyond some error, so I doubt it makes a difference.