@HurricanKai 2020-10-20
Having both an overload for List & IROList
This is only useful in the WithTrivia case, so there the code I chose was
```cs
null => null,
List<TFrom> list => GreenNode.CreateList(list, select),
IReadOnlyList list => GreenNode.CreateList(list, select),
_ => GreenNode.CreateList(enumerable.ToList(), select)
```