@HurricanKai 2020-10-21
I've updated the original post to show current numbers.